perf: big refactor + smaller size
