+++ src/internal/common.ts
@@ -1,0 +1,7 @@
+import type { State } from './types.ts'
+
+export const newInstance = (name: string, args: string | number = ``): string =>
+  `new ${name}${args === `` ? `` : `(${args})`}`
+
+export const bindingName = (value: object, state: State): string =>
+  state._bindings.get(value)!._name
